fix task serialization
